Enhance UI and functionality of SelectableIconButton and tool grid
- Added shortcut labels to SelectableIconButton for improved user experience.
- Refactored tool grid buttons to utilize shortcut labels and updated tooltips for clarity.
- Improved text placement calculations in text tools for better alignment.
- Optimized map selector to sort available and out-of-rotation maps for enhanced usability.

+++ lib/const/drawing_element.dart
@@ -155,7 +155,8 @@ class FreeDrawing extends DrawingElement with HiveObjectMixin {
       final path = Path();
       final screenPoint = coordinateSystem.coordinateToScreen(listOfPoints[0]);
       final dotRadius = (coordinateSystem.scale(Settings.brushSize * 0.25))
-          .clamp(1.0, coordinateSystem.scale(2.0));
+          .clamp(1.0, coordinateSystem.scale(2.0))
+          .toDouble();
 
       path.addOval(Rect.fromCircle(center: screenPoint, radius: dotRadius));
       _path = path;
